+++ satkit/extensions_to_external_classes/base_model_extensions.py
