+++ satkit/extensions_to_external_classes/enum_extensions.py
